Clean-up example/predicat-tree
